fix(avif): clippy --tests lints in test code (bad bit mask, redundant cast, range loop)
The CI clippy step runs with --tests + -D warnings; these test-only lints were
invisible to a plain `cargo clippy` (which skips #[cfg(test)] targets). Drop two
tautological `(x & 0x3f) <= 63` asserts (kept the meaningful dc-sign check),
remove a redundant i32->i32 cast, and index maxdiff via iter_mut().enumerate().

+++ crates/core/src/raster/avif/itx.rs
@@ -1011,7 +1011,7 @@ mod tests {
     fn inv_txfm_2d_full_runs() {
         // Full 8×8 DCT_DCT path with a spread of coefficients: must run and
         // produce a finite residual of the right size.
-        let mut cf: Vec<i32> = (0..64).map(|i| ((i * 11) % 23) as i32 - 11).collect();
+        let mut cf: Vec<i32> = (0..64).map(|i| ((i * 11) % 23) - 11).collect();
         let res = inv_txfm_residual(&mut cf, 1, txtp::DCT_DCT, 20);
         assert_eq!(res.len(), 64);
         assert!(res.iter().all(|&v| v.abs() < 1 << 20));
